build: move packaging out of gulp folder
* Moves the packaging out of the `tools/gulp` folder. This change is in favor of exposing the build-tools to Github (for use in flex-layout)
* Package tools no longer depend on the `task_helpers.ts` file from `tools/gulp` (Necessary for publishing)
* Relative imports to the packaging have been replaced with `material2-build-tools` (Useful for sharing existing gulp tasks with flex-layout)
