+++ gulpfile.js
@@ -6,9 +6,20 @@
 
 const path = require('path');
 
+const tsconfigPath = path.join(__dirname, 'tools/gulp/tsconfig.json');
+const tsconfig = require(tsconfigPath);
+
 // Register TS compilation.
 require('ts-node').register({
-  project: path.join(__dirname, 'tools/gulp/tsconfig.json')
+  project: tsconfigPath
+});
+
+// The gulp tsconfig file maps specific imports to relative paths. In combination with ts-node
+// this doesn't work because the JavaScript output will still refer to the imports instead of
+// to the relative path. Tsconfig-paths can be used to support path mapping inside of Node.
+require("tsconfig-paths").register({
+  baseUrl: path.dirname(tsconfigPath),
+  paths: tsconfig.compilerOptions.paths
 });
 
 require('./tools/gulp/gulpfile');
