+++ tools/gulp/tasks/coverage.ts
@@ -4,7 +4,7 @@ import {existsSync} from 'fs-extra';
 import {spawnSync} from 'child_process';
 import {isTravisMasterBuild} from '../util/travis-ci';
 import {openFirebaseDashboardApp} from '../util/firebase';
-import {buildConfig} from '../packaging/build-config';
+import {buildConfig} from 'material2-build-tools';
 
 /** Path to the file that includes all coverage information form Karma. */
 const coverageResultFile = join(buildConfig.outputDir, 'coverage/coverage-summary.json');
