+++ tools/gulp/tasks/docs.ts
@@ -1,7 +1,7 @@
 import {task, src, dest} from 'gulp';
 import {Dgeni} from 'dgeni';
 import * as path from 'path';
-import {buildConfig} from '../packaging/build-config';
+import {buildConfig} from 'material2-build-tools';
 
 // There are no type definitions available for these imports.
 const markdown = require('gulp-markdown');
