+++ tools/gulp/tasks/publish.ts
@@ -2,8 +2,8 @@ import {spawn} from 'child_process';
 import {existsSync, statSync} from 'fs-extra';
 import {join} from 'path';
 import {task} from 'gulp';
-import {execTask, sequenceTask} from '../util/task_helpers';
-import {buildConfig} from '../packaging/build-config';
+import {execTask} from '../util/task_helpers';
+import {buildConfig, sequenceTask} from 'material2-build-tools';
 import {yellow, green, red, grey} from 'chalk';
 import * as minimist from 'minimist';
