+++ tools/gulp/tasks/validate-release.ts
@@ -2,10 +2,9 @@ import {task} from 'gulp';
 import {readFileSync, existsSync} from 'fs';
 import {join} from 'path';
 import {green, red} from 'chalk';
-import {sequenceTask} from '../util/task_helpers';
 import {releasePackages} from './publish';
 import {sync as glob} from 'glob';
-import {buildConfig} from '../packaging/build-config';
+import {buildConfig, sequenceTask} from 'material2-build-tools';
 
 /** Path to the directory where all releases are created. */
 const releasesDir = join(buildConfig.outputDir, 'releases');
