+++ tools/gulp/tsconfig.json
@@ -1,5 +1,3 @@
-// TypeScript config file that will be used to compile the gulp tasks. The normal gulpfile will
-// use TS-Node to compile the gulp tasks at runtime.
 {
   "compilerOptions": {
     "experimentalDecorators": true,
@@ -13,7 +11,11 @@
     "target": "es5",
     "types": [
       "node"
-    ]
+    ],
+    "baseUrl": ".",
+    "paths": {
+      "material2-build-tools": ["../package-tools/index"]
+    }
   },
   "files": [
     "gulpfile.ts"
