+++ tools/gulp/util/task_helpers.ts
@@ -2,16 +2,12 @@ import * as child_process from 'child_process';
 import * as fs from 'fs';
 import * as gulp from 'gulp';
 import * as path from 'path';
-import {yellow} from 'chalk';
-import {buildConfig} from '../packaging/build-config';
+import {buildConfig} from 'material2-build-tools';
 
 /* Those imports lack typings. */
 const gulpClean = require('gulp-clean');
 const gulpRunSequence = require('run-sequence');
-const gulpSass = require('gulp-sass');
 const gulpConnect = require('gulp-connect');
-const gulpIf = require('gulp-if');
-const gulpCleanCss = require('gulp-clean-css');
 
 // There are no type definitions available for these imports.
 const resolveBin = require('resolve-bin');
@@ -43,17 +39,6 @@ export function tsBuildTask(tsConfigPath: string) {
 export function ngcBuildTask(tsConfigPath: string) {
   return execNodeTask('@angular/compiler-cli', 'ngc', ['-p', tsConfigPath]);
 }
-
-/** Create a SASS Build Task. */
-export function sassBuildTask(dest: string, root: string, minify = false) {
-  return () => {
-    return gulp.src(_globify(root, '**/*.scss'))
-      .pipe(gulpSass().on('error', gulpSass.logError))
-      .pipe(gulpIf(minify, gulpCleanCss()))
-      .pipe(gulp.dest(dest));
-  };
-}
-
 
 /** Options that can be passed to execTask or execNodeTask. */
 export interface ExecTaskOptions {
@@ -175,22 +160,5 @@ export function serverTask(packagePath: string, livereload = true) {
         ])];
       }
     });
-  };
-}
-
-/** Triggers a reload when livereload is enabled and a gulp-connect server is running. */
-export function triggerLivereload() {
-  console.log(yellow('Server: Changes were detected and a livereload was triggered.'));
-  return gulp.src('dist').pipe(gulpConnect.reload());
-}
-
-
-/** Create a task that's a sequence of other tasks. */
-export function sequenceTask(...args: any[]) {
-  return (done: any) => {
-    gulpRunSequence(
-      ...args,
-      done
-    );
   };
 }
