+++ tools/package-tools/copy-files.ts
