+++ tools/package-tools/find-build-config.ts
