+++ tools/package-tools/gulp/sequence-task.ts
@@ -1,0 +1,12 @@
+// This import does not have any type definitions.
+const gulpRunSequence = require('run-sequence');
+
+/** Create a task that's a sequence of other tasks. */
+export function sequenceTask(...args: any[]) {
+  return (done: any) => {
+    gulpRunSequence(
+      ...args,
+      done
+    );
+  };
+}
