+++ tools/package-tools/inline-resources.ts
