+++ tools/package-tools/metadata-inlining.ts
