+++ tools/package-tools/metadata-reexport.ts
