+++ tools/package-tools/minify-sources.ts
